Refactor: API 모듈화 처리
